+++ src/datasources/asset/types/types.ts
@@ -1,7 +1,6 @@
 import { DataSourceJsonData } from "@grafana/data";
-import { AssetSummaryQuery } from "./AssetSummaryQuery.types";
-import { CalibrationForecastQuery } from "./CalibrationForecastQuery.types";
-import { ListAssetsQuery } from "./ListAssets.types";
+import { DataQuery } from "@grafana/schema";
+
 
 export enum AssetQueryType {
   None = "",
@@ -10,7 +9,9 @@ export enum AssetQueryType {
   AssetSummary = "Asset Summary"
 }
 
-export type AssetQuery = ListAssetsQuery | CalibrationForecastQuery | AssetSummaryQuery;
+export interface AssetQuery extends DataQuery {
+  type: AssetQueryType
+}
 
 export interface AssetFeatureToggles {
   calibrationForecast: boolean;