+++ src/datasources/tag/module.ts
@@ -1,8 +1,9 @@
 import { DataSourcePlugin } from '@grafana/data';
 import { TagDataSource } from './TagDataSource';
 import { TagQueryEditor } from './components/TagQueryEditor';
-import { HttpConfigEditor } from 'core/components/HttpConfigEditor';
+import { TagConfigEditor } from "./TagConfigEditor";
+import { TagDataSourceOptions, TagQuery } from "./types";
 
-export const plugin = new DataSourcePlugin(TagDataSource)
-  .setConfigEditor(HttpConfigEditor)
+export const plugin = new DataSourcePlugin<TagDataSource, TagQuery, TagDataSourceOptions>(TagDataSource)
+  .setConfigEditor(TagConfigEditor)
   .setQueryEditor(TagQueryEditor);